--- tests/mercury.py
@@ -1,34 +1,0 @@
-import unittest
-
-from utils.channel_access import ChannelAccess
-from utils.ioc_launcher import get_default_ioc_dir, EPICS_TOP
-from utils.test_modes import TestModes
-from utils.testing import get_running_lewis_and_ioc, skip_if_recsim
-import os
-
-
-DEVICE_PREFIX = "MERCURY_01"
-
-IOCS = [
-    {
-        "name": DEVICE_PREFIX,
-        "directory": os.path.join(EPICS_TOP, "ioc", "master", "MERCURY_ITC", "iocBoot", "iocMERCURY-IOC-01"),
-        "macros": {},
-        "emulator": "Mercury",
-    },
-]
-
-
-TEST_MODES = [TestModes.RECSIM, TestModes.DEVSIM]
-
-
-class MercuryTests(unittest.TestCase):
-    """
-    Tests for the Mercury IOC.
-    """
-    def setUp(self):
-        self._lewis, self._ioc = get_running_lewis_and_ioc("Mercury", DEVICE_PREFIX)
-        self.ca = ChannelAccess(device_prefix=DEVICE_PREFIX)
-
-    def test_WHEN_ioc_is_started_THEN_ioc_is_not_disabled(self):
-        self.ca.assert_that_pv_is("DISABLE", "COMMS ENABLED")
